requirements.txt
python-whois
pyyaml
schedule